LateralNav -> Toolbar && created toolbarUtils.js
